factor parentframe export away from uitoggle. Rough in plugin generation code Add parentframe page (not working cause depends on frames not converted yet)
