+++ src/Pages/parentframe.html
@@ -4,48 +4,10 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=Edge" />
     <title>MessageHeaderAnalyzer</title>
-
-    <link rel="stylesheet" href="https://static2.sharepointonline.com/files/fabric/office-ui-fabric-js/1.5.0/css/fabric.min.css">
-    <link rel="stylesheet" href="https://static2.sharepointonline.com/files/fabric/office-ui-fabric-js/1.5.0/css/fabric.components.min.css">
-    <link rel="stylesheet" type="text/css" href="../Content/%version%/uiToggle.css" />
-
-    <script crossorigin="anonymous" src="https://ajax.aspnetcdn.com/ajax/jQuery/jquery-3.5.1.min.js"></script>
-    <script>
-        // Fallback to load jQuery from a local path if the CDN is unavailable.
-        (window.jQuery || document.write('<script src="../node_modules/jquery/dist/jquery.min.js"><\/script>'));
-    </script>
-
-    <script crossorigin="anonymous" src="https://cdnjs.cloudflare.com/ajax/libs/stacktrace.js/2.0.2/stacktrace-with-promises-and-json-polyfills.min.js"></script>
-    <script>
-        (window.StackTrace || document.write('<script src="../node_modules/stacktrace-js/dist/stacktrace-with-promises-and-json-polyfills.min.js"><\/script>'));
-    </script>
-
-    <script src="../Scripts/%version%/diag.min.js"></script>
-
-    <script crossorigin="anonymous" src="https://ajax.aspnetcdn.com/ajax/4.0/MicrosoftAjax.js"></script>
+    <!--<script crossorigin="anonymous" src="https://ajax.aspnetcdn.com/ajax/4.0/MicrosoftAjax.js"></script>
     <script>
         (window.Sys || document.write('<script src="../packages/MicrosoftAjax.4.0.20526.0/Content/Scripts/MicrosoftAjax.js"><\/script>'));
-    </script>
-
-    <script crossorigin="anonymous" src="https://appsforoffice.microsoft.com/lib/1/hosted/office.debug.js"></script>
-    <script>
-        (window.Office || document.write('<script src="../node_modules/@microsoft/office-js/dist/office.debug.js"><\/script>'));
-    </script>
-
-    <script crossorigin="anonymous" src="https://static2.sharepointonline.com/files/fabric/office-ui-fabric-js/1.5.0/js/fabric.min.js"></script>
-    <script>
-        (window.fabric || document.write('<script src="../node_modules/office-ui-fabric-js/dist/js/fabric.min.js"><\/script>'));
-    </script>
-
-    <script src="../node_modules/jwt-decode/build/jwt-decode.js"></script>
-    <script src="../Scripts/%version%/Strings.min.js"></script>
-    <script src="../Scripts/%version%/Dates.min.js"></script>
-    <script src="../Scripts/%version%/poster.min.js"></script>
-    <script src="../Scripts/%version%/Errors.min.js"></script>
-    <script src="../Scripts/%version%/GetHeaders.min.js"></script>
-    <script src="../Scripts/%version%/GetHeadersEWS.min.js"></script>
-    <script src="../Scripts/%version%/GetHeadersRest.min.js"></script>
-    <script src="../Scripts/%version%/uiToggle.min.js"></script>
+    </script>-->
 </head>
 <body class="ms-Fabric">
     <div id="uiToggleFrame">
